Refactor tests so that they use the actual unlock functionality
