+++ rotkehlchen/tests/api/test_messages.py
@@ -5,7 +5,7 @@ from rotkehlchen.tests.utils.api import api_url_for, assert_proper_response
 from rotkehlchen.tests.utils.history import mock_history_processing_and_exchanges
 
 
-@pytest.mark.parametrize('added_exchanges', [('poloniex')])
+@pytest.mark.parametrize('added_exchanges', [('poloniex',)])
 def test_query_messages(rotkehlchen_api_server_with_exchanges):
     """Test that querying the messages endpoint returns notifications for the user"""
     rotki = rotkehlchen_api_server_with_exchanges.rest_api.rotkehlchen
