+++ rotkehlchen/tests/fixtures/exchanges/bitmex.py
@@ -1,6 +1,7 @@
 import pytest
 
 from rotkehlchen.exchanges.bitmex import Bitmex
+from rotkehlchen.tests.utils.exchanges import create_test_bitmex
 from rotkehlchen.user_messages import MessagesAggregator
 
 TEST_BITMEX_API_KEY = 'XY98JYVL15Zn-iU9f7OsJeVf'
@@ -8,22 +9,21 @@ TEST_BITMEX_API_SECRET = b'671tM6f64bt6KhteDakj2uCCNBt7HhZVEE7H5x16Oy4zb1ag'
 
 
 @pytest.fixture
-def mock_bitmex(database, inquirer):  # pylint: disable=unused-argument
+def mock_bitmex(
+        database,
+        inquirer,  # pylint: disable=unused-argument
+        function_scope_messages_aggregator,
+):
     # API key/secret from tests cases here: https://www.bitmex.com/app/apiKeysUsage
-    bitmex = Bitmex(
-        api_key='LAqUlngMIQkIUjXMUreyu3qn',
-        secret=b'chNOOS4KvNXR_Xq4k4c9qsfoKWvnDecLATCRlcBwyKDYnWgO',
+    bitmex = create_test_bitmex(
         database=database,
-        msg_aggregator=MessagesAggregator(),
+        msg_aggregator=function_scope_messages_aggregator,
     )
-
-    bitmex.first_connection_made = True
     return bitmex
 
 
 @pytest.fixture
-def test_bitmex(database, inquirer):  # pylint: disable=unused-argument
-    # API key/secret from tests cases here: https://www.bitmex.com/app/apiKeysUsage
+def sandbox_bitmex(database, inquirer):  # pylint: disable=unused-argument
     bitmex = Bitmex(
         api_key=TEST_BITMEX_API_KEY,
         secret=TEST_BITMEX_API_SECRET,
