+++ rotkehlchen/tests/fixtures/exchanges/poloniex.py
@@ -1,11 +1,6 @@
 import pytest
 
-from rotkehlchen.exchanges.poloniex import Poloniex
-from rotkehlchen.tests.utils.factories import make_api_key, make_api_secret
-
-
-class MockPoloniex(Poloniex):
-    pass
+from rotkehlchen.tests.utils.exchanges import create_test_poloniex
 
 
 @pytest.fixture(scope='session')
@@ -14,13 +9,10 @@ def poloniex(
         session_inquirer,  # pylint: disable=unused-argument
         messages_aggregator,
 ):
-    mock = MockPoloniex(
-        api_key=make_api_key(),
-        secret=make_api_secret(),
+    return create_test_poloniex(
         database=session_database,
         msg_aggregator=messages_aggregator,
     )
-    return mock
 
 
 @pytest.fixture(scope='function')
@@ -29,10 +21,7 @@ def function_scope_poloniex(
         inquirer,  # pylint: disable=unused-argument
         function_scope_messages_aggregator,
 ):
-    mock = MockPoloniex(
-        api_key=make_api_key(),
-        secret=make_api_secret(),
+    return create_test_poloniex(
         database=database,
         msg_aggregator=function_scope_messages_aggregator,
     )
-    return mock
