+++ rotkehlchen/tests/fixtures/history.py
@@ -3,15 +3,15 @@ import pytest
 from rotkehlchen.db.dbhandler import DBHandler
 from rotkehlchen.exchanges.manager import ExchangeManager
 from rotkehlchen.externalapis.cryptocompare import Cryptocompare
-from rotkehlchen.fval import FVal
 from rotkehlchen.history import PriceHistorian, TradesHistorian
+from rotkehlchen.tests.utils.history import maybe_mock_price_queries
 
 TEST_HISTORY_DATA_START = "01/01/2015"
 
 
 @pytest.fixture
-def cryptocompare(accounting_data_dir, database):
-    return Cryptocompare(data_directory=accounting_data_dir, database=database)
+def cryptocompare(data_dir, database):
+    return Cryptocompare(data_directory=data_dir, database=database)
 
 
 @pytest.fixture(scope='session')
@@ -21,7 +21,7 @@ def session_cryptocompare(session_data_dir, session_database):
 
 @pytest.fixture
 def price_historian(
-        accounting_data_dir,
+        data_dir,
         inquirer,  # pylint: disable=unused-argument
         should_mock_price_queries,
         mocked_price_queries,
@@ -31,36 +31,25 @@ def price_historian(
     # is called make sure its instance is always starting from scratch
     PriceHistorian._PriceHistorian__instance = None
     historian = PriceHistorian(
-        data_directory=accounting_data_dir,
+        data_directory=data_dir,
         history_date_start=TEST_HISTORY_DATA_START,
         cryptocompare=cryptocompare,
     )
-    if should_mock_price_queries:
-        def mock_historical_price_query(from_asset, to_asset, timestamp):
-            if from_asset == to_asset:
-                return FVal(1)
-
-            try:
-                price = mocked_price_queries[from_asset.identifier][to_asset.identifier][timestamp]
-            except KeyError:
-                raise AssertionError(
-                    f'No mocked price found from {from_asset.identifier} to '
-                    f'{to_asset.identifier} at {timestamp}',
-                )
-
-            return price
-
-        historian.query_historical_price = mock_historical_price_query
+    maybe_mock_price_queries(
+        historian=historian,
+        should_mock_price_queries=should_mock_price_queries,
+        mocked_price_queries=mocked_price_queries,
+    )
 
     return historian
 
 
 @pytest.fixture
-def trades_historian(accounting_data_dir, function_scope_messages_aggregator, blockchain):
-    database = DBHandler(accounting_data_dir, '123', function_scope_messages_aggregator)
+def trades_historian(data_dir, function_scope_messages_aggregator, blockchain):
+    database = DBHandler(data_dir, '123', function_scope_messages_aggregator)
     exchange_manager = ExchangeManager(msg_aggregator=function_scope_messages_aggregator)
     historian = TradesHistorian(
-        user_directory=accounting_data_dir,
+        user_directory=data_dir,
         db=database,
         msg_aggregator=function_scope_messages_aggregator,
         exchange_manager=exchange_manager,
